Places creation page done
