+++ client/src/App.jsx
@@ -3,9 +3,11 @@ import {Route, Routes} from "react-router-dom";
 import IndexPage from './pages/IndexPage';
 import LoginPage from './pages/LoginPage';
 import RegisterPage from './pages/RegisterPage';
-import AccountPage from './pages/AccountPage';
-import Layout from './Layout';
+import ProfilePage from './pages/ProfilePage';
+import PlacesPage from './pages/PlacesPage';
+import PlacesFormPage from './pages/PlacesFormPage';
 import { UserContextProvider } from './UserContext';
+import Layout from './Layout';
 import axios from 'axios';
 
 axios.defaults.baseURL = 'http://127.0.0.1:4000';
@@ -20,8 +22,10 @@ function App() {
           <Route index element={<IndexPage />} />
           <Route path="/login" element={<LoginPage />} />
           <Route path="/register" element={<RegisterPage />} />
-          <Route path="/account/:subpage?" element={<AccountPage />} />
-          <Route path="/account/:subpage/:action" element={<AccountPage />} />
+          <Route path="/account/" element={<ProfilePage />} />
+          <Route path="/account/places" element={<PlacesPage />} />
+          
+          <Route path="/account/places/new" element={<PlacesFormPage />} />
         </Route>
       </Routes>
     </UserContextProvider>
